Add invoked method params and info to summary. Add edgeTargetCount and edgeSourceCount to Node. Log failure to create or find doc items.

+++ src/Stratis.DevEx.Drawing/VisJS/Network.cs
@@ -163,7 +163,7 @@ namespace Stratis.DevEx.Drawing
         public NetworkFont? Font { get; set; }
 
         [JsonProperty("mass"), JsonSuppressDefaultValue]
-        public int? Mass { get; set; }
+        public double? Mass { get; set; }
     }
 
     public class NetworkOptions
